refactor(products): update routes to include organization
