+++ api/v1/routes/product.py
@@ -24,26 +24,157 @@ from api.v1.schemas.product import (
     ProductCategoryRetrieve,
     ProductCommentCreate,
     ProductCommentsSchema,
+    ProductDetail,
 )
 from api.utils.dependencies import get_current_user
 from api.v1.services.user import user_service
 from api.v1.models import User
 
-product = APIRouter(prefix="/products", tags=["Products"])
+superadmin_product = APIRouter(prefix="/products", tags=["Products"])
 
 
-@product.get("", response_model=success_response, status_code=200)
+@superadmin_product.get("", response_model=success_response, status_code=200)
 async def get_all_products(
     current_user: Annotated[User, Depends(user_service.get_current_super_admin)],
-    limit: Annotated[int, Query(
-        ge=1, description="Number of products per page")] = 10,
-    skip: Annotated[int, Query(
-        ge=1, description="Page number (starts from 1)")] = 0,
+    limit: Annotated[int, Query(ge=1, description="Number of products per page")] = 10,
+    skip: Annotated[int, Query(ge=1, description="Page number (starts from 1)")] = 0,
     db: Session = Depends(get_db),
 ):
     """Endpoint to get all products. Only accessible to superadmin"""
 
     return paginated_response(db=db, model=Product, limit=limit, skip=skip)
+
+
+product = APIRouter(prefix="/organizations/{org_id}/products", tags=["Products"])
+
+
+@product.post("", status_code=status.HTTP_201_CREATED)
+def product_create(
+    org_id: str,
+    product: ProductCreate,
+    current_user: Annotated[User, Depends(user_service.get_current_user)],
+    db: Session = Depends(get_db),
+):
+    created_product = product_service.create(
+        db=db, schema=product, org_id=org_id, current_user=current_user
+    )
+
+    return success_response(
+        status_code=status.HTTP_201_CREATED,
+        message="Product created successfully",
+        data=jsonable_encoder(created_product),
+    )
+
+
+@product.get(
+    "/{product_id}",
+    response_model=dict[str, int | str | bool | ProductDetail],
+    summary="Get product detail",
+    description="Endpoint to get detail about the product with the given `id`",
+)
+async def get_product_detail(
+    org_id: str,
+    product_id: str,
+    db: Session = Depends(get_db),
+    current_user: User = Depends(user_service.get_current_user),
+):
+    """
+    Retrieve product detail
+
+    This endpoint retrieve details about a product
+
+    Args:
+        org_id (UUID): The unique identifier of the organization
+        product_id (UUID): The unique identifier of the product to be retrieved.
+        db (Session): The database session, provided by the `get_db` dependency.
+        current_user (User): The currently authenticated user, obtained from the `get_current_user` dependency.
+
+    Returns:
+        ProductDetail: The detail of the product matching the given id
+
+    Raises:
+        HTTPException: If the product with the specified `id` does not exist, a 404 error is raised.
+    """
+
+    product = product_service.fetch_single_by_organization(
+        db, org_id, product_id, current_user
+    )
+
+    return {
+        "status_code": status.HTTP_200_OK,
+        "success": True,
+        "message": "Product fetched successfully",
+        "data": product,
+    }
+
+
+@product.put("/{id}", response_model=ResponseModel)
+async def update_product(
+    id: str,
+    product_update: ProductUpdate,
+    current_user: User = Depends(get_current_user),
+    db: Session = Depends(get_db),
+):
+    """
+    Update the details of an existing product.
+
+    This endpoint updates a product's attributes such as name, price, description, and tag.
+    It ensures that the product exists before performing the update. The `updated_at` timestamp
+    is set to the current time to reflect when the update occurred.
+
+    Args:
+        id (UUID): The unique identifier of the product to be updated.
+        product (ProductUpdate): The new product data to be updated.
+        current_user (User): The currently authenticated user, obtained from the `get_current_user` dependency.
+        db (Session): The database session, provided by the `get_db` dependency.
+
+    Returns:
+        ProductUpdate: The updated product details.
+
+    Raises:
+        HTTPException: If the product with the specified `id` does not exist, a 404 error is raised.
+
+    Example:
+        PUT /product/123e4567-e89b-12d3-a456-426614174000
+        {
+            "name": "New Product Name",
+            "price": 29.99,
+            "description": "Updated description",
+        }
+    """
+
+    updated_product = product_service.update(db, id=str(id), schema=product_update)
+
+    # Prepare the response
+    return success_response(
+        status_code=200,
+        message="Product updated successfully",
+        data=jsonable_encoder(updated_product),
+    )
+
+
+@product.delete("/{product_id}", status_code=status.HTTP_204_NO_CONTENT)
+def delete_product(
+    org_id: str,
+    product_id: str,
+    current_user: User = Depends(user_service.get_current_user),
+    db: Session = Depends(get_db),
+):
+    """Enpoint to delete a product
+
+    Args:
+        product_id (str): The unique identifier of the product to be deleted
+        current_user (User): The currently authenticated user, obtained from the `get_current_user` dependency.
+        db (Session): The database session, provided by the `get_db` dependency.
+
+    Raises:
+        HTTPException: 401 FORBIDDEN (Current user is not a authenticated)
+        HTTPException: 404 NOT FOUND (Product to be deleted cannot be found)
+    """
+
+    product_service.delete(
+        db=db, org_id=org_id, product_id=product_id, current_user=current_user
+    )
 
 
 @product.get(
@@ -63,8 +194,7 @@ async def get_products_by_filter_status(
             message="Products retrieved successfully", status_code=200, data=products
         )
     except Exception as e:
-        raise HTTPException(
-            status_code=500, detail="Failed to retrieve products")
+        raise HTTPException(status_code=500, detail="Failed to retrieve products")
 
 
 @product.get(
@@ -84,14 +214,13 @@ async def get_products_by_status(
             message="Products retrieved successfully", status_code=200, data=products
         )
     except Exception as e:
-        raise HTTPException(
-            status_code=500, detail="Failed to retrieve products")
+        raise HTTPException(status_code=500, detail="Failed to retrieve products")
 
 
 @product.get("/categories", response_model=success_response, status_code=200)
 def retrieve_categories(
     db: Session = Depends(get_db),
-    current_user: User = Depends(user_service.get_current_user)
+    current_user: User = Depends(user_service.get_current_user),
 ):
     """
     Retrieve all product categories from database
@@ -100,31 +229,29 @@ def retrieve_categories(
     categories = ProductCategoryService.fetch_all(db)
 
     categories_filtered = list(
-        map(lambda x: ProductCategoryRetrieve.model_validate(x), categories))
+        map(lambda x: ProductCategoryRetrieve.model_validate(x), categories)
+    )
 
-    if (len(categories_filtered) == 0):
+    if len(categories_filtered) == 0:
         categories_filtered = [{}]
 
     return success_response(
         message="Categories retrieved successfully",
         status_code=200,
-        data=jsonable_encoder(categories_filtered)
+        data=jsonable_encoder(categories_filtered),
     )
 
 
-@product.get("/{org_id}", status_code=status.HTTP_200_OK, response_model=ProductList)
 @product.get(
-    "/organizations/{org_id}",
+    "",
     status_code=status.HTTP_200_OK,
     response_model=ProductList,
 )
 def get_organization_products(
     org_id: str,
     current_user: Annotated[User, Depends(user_service.get_current_user)],
-    limit: Annotated[int, Query(
-        ge=1, description="Number of products per page")] = 10,
-    page: Annotated[int, Query(
-        ge=1, description="Page number (starts from 1)")] = 1,
+    limit: Annotated[int, Query(ge=1, description="Number of products per page")] = 10,
+    page: Annotated[int, Query(ge=1, description="Page number (starts from 1)")] = 1,
     db: Session = Depends(get_db),
 ):
     """
@@ -171,7 +298,7 @@ def get_organization_products(
 async def get_product_stock(
     id: str,
     current_user: Annotated[User, Depends(user_service.get_current_user)],
-    db: Session = Depends(get_db)
+    db: Session = Depends(get_db),
 ):
     """
     Retrieve the current stock level for a specific product.
@@ -198,52 +325,7 @@ async def get_product_stock(
     )
 
 
-@product.put("/{id}", response_model=ResponseModel)
-async def update_product(
-    id: str,
-    product_update: ProductUpdate,
-    current_user: User = Depends(get_current_user),
-    db: Session = Depends(get_db),
-):
-    """
-    Update the details of an existing product.
-
-    This endpoint updates a product's attributes such as name, price, description, and tag.
-    It ensures that the product exists before performing the update. The `updated_at` timestamp
-    is set to the current time to reflect when the update occurred.
-
-    Args:
-        id (UUID): The unique identifier of the product to be updated.
-        product (ProductUpdate): The new product data to be updated.
-        current_user (User): The currently authenticated user, obtained from the `get_current_user` dependency.
-        db (Session): The database session, provided by the `get_db` dependency.
-
-    Returns:
-        ProductUpdate: The updated product details.
-
-    Raises:
-        HTTPException: If the product with the specified `id` does not exist, a 404 error is raised.
-
-    Example:
-        PUT /product/123e4567-e89b-12d3-a456-426614174000
-        {
-            "name": "New Product Name",
-            "price": 29.99,
-            "description": "Updated description",
-        }
-    """
-
-    updated_product = product_service.update(
-        db, id=str(id), schema=product_update)
-
-    # Prepare the response
-    return success_response(
-        status_code=200,
-        message="Product updated successfully",
-        data=jsonable_encoder(updated_product),
-    )
-
-@product.post('/categories/{org_id}', status_code=status.HTTP_201_CREATED)
+@product.post("/categories", status_code=status.HTTP_201_CREATED)
 def create_product_category(
     org_id: str,
     category_schema: ProductCategoryCreate,
@@ -264,7 +346,9 @@ def create_product_category(
         HTTPException: 401 FORBIDDEN (Current user is not a authenticated)
     """
 
-    new_category = ProductCategoryService.create(db, org_id, category_schema, current_user)
+    new_category = ProductCategoryService.create(
+        db, org_id, category_schema, current_user
+    )
 
     return success_response(
         status_code=status.HTTP_201_CREATED,
@@ -272,18 +356,20 @@ def create_product_category(
         data=jsonable_encoder(new_category),
     )
 
-@product.post("/{product_id}/comments", status_code=status.HTTP_201_CREATED, response_model=ProductCommentsSchema)
+
+@product.post(
+    "/{product_id}/comments",
+    status_code=status.HTTP_201_CREATED,
+    response_model=ProductCommentsSchema,
+)
 def create_product_comment(
     product_id: str,
     comment: ProductCommentCreate,
     current_user: User = Depends(user_service.get_current_user),
-    db: Session = Depends(get_db)
+    db: Session = Depends(get_db),
 ):
     product_comment = product_comment_service.create(
-        db,
-        comment,
-        current_user.id,
-        product_id
+        db, comment, current_user.id, product_id
     )
     return success_response(
         status_code=status.HTTP_201_CREATED,
